create cli #2
created the start point of the cli. got authorization working for
logging in a user and some checks on return values to not constantly
error out for things that can be repeated. will start to separate out
commands more and figure out how best to write out the interactive
portion of it. maybe configure it in a way that can be interactive and
just executed once via the command line.

have to add in some kind of mechanism for storing the session ids for
longer.