cli storage interaction #5
added in the ability to create/update storage mediums for a given
server. fairly straight forward in terms of features. still need to
figure out what to do about deleting but next will be working on fs.

 - updated server code to fix bugs and sql
 - small changes to schema and common json wrappers

may start working on a common list of errors that the server can return
so that it can be shared between whatever wants it vs relying upon
strings that would need to be individually updated.