Totp update and patch method #27
made updates to the Totp struct so that it can be updated in the
database. uses a similar pattern as the Hash class with the
HistoryField.
added in the ability to call a patch endpoint to update the Totp with
new data or to just regen the secret. the secret will always be
regenerated if any of the other values are set.