config process update #23
the configuration process was not the best (still may not be) so I went
through and updated it to move away from the builder processes and just
use static functions to create the state for the server. since this
includes changes not staged this will not build but it is also not
relevant to the secrets management. should have done this differently
but here we are.