moved kms into separate repo #23
moved all of the key management code into a separate repository and then
updated the server accordingly. it did require a few changes to the to
the kms repo before the server would successfully build. the kms code is
different than what was implemented originally but still effectively the
same.