updated session handling #23
updated logic to use the new session keys manager. this will deviate
from the original ticket as the sessions do not use versions and is just
a list of previous values in order. you can only removed the oldest
values from the list.

also added in the version logic for password secrets.