base rework #41
the base changes made for getting the new pepper rotation setup for the
server. it required reworking a few things to work encryption but was
not too bad. trying out puting more logic into the Password struct to
handle the various processes. also created a PeppersManager struct to
handle simple things for accessing the peppers in memory. still needs to
be tested but this is already getting big.