bug fix and cli #41
the pepper rotation process when deleting went fairly smooth and did not
require too many fixes. had to add in a way to update and delete
passwords for the current user since the cli did not handle that yet.
some other small changes that were noticed when working on this
particular commit.