storage update #53
updated the logic and moved some code around. the storage code has been
moved into fs and some of the structs have been renamed. now instead of
types there are backends and a storage is configured with a specific
backend. the same general idea still applies just with different naming.
all of the files/directories that are under a certain storage backend
with have a node (maybe not the best name but whatever) that stores
relevant information for that specific item. the storage id will no
longer be stored in the node and will just be a field in the table that
can be more easily referenced.

after going over a bunch of the code for storage objects, they are very
similar to root items and on the fence about how best to store the data.
for now going to stick with what is currently setup.

need to look into better handling the pathing situation has it is based
on std::path and that changes depending on the system that it is running
on. the backend data will keep using it as it will require it but the
database does not care and the code needs to better reflect that.