cli update and fixes #53
modified the output of some fs commands to have better output and made
the individual records output more consistent between commands. moved
the storage commands to the new location and also updated the root fs
item to contain the name of the storage interface so that the root can
have more of an indication of the name vs just having nothing.