path and fs updates #53
main changes is changing the path type for file, directory, and root to
now be a regular string vs a path. the file system path information will
be stored in the backend data and the path for the record will be
aggnostic from the file system. now if desired the file system path can
be separate from the database path.

updated more storage logic to better handle getting the data necessary
for creating new fs items. took some thinking and ended on the current
idea to try converting the fs item into the necessary parts if it is a
container type otherwise return the fs item.