cli download #52
added in the ability for the cli to download a file from the server. not
a whole lot of additional features attached to it but it will work for
now. it will calculate the hash against the provided checksum and notify
the user if the hashes do not match. also some small info about how the
download when, eg. speed, time, and total bytes.