session cleanup logic #49
added in the logic for cleaning up sessions from the database. also
updated the logic for running jobs and saving results to a file in the
servers data directory. jobs will now run if they have never been run
before. this migh change into an option of some kind so that some jobs
that don't need to run if they never have been don't have to run.