updated config process
no ticket for this. the previous process had been bothering me for a while
because it always seems convoluted and there was logic that was happening twice
for no good reason. the newer process follows the same one but cleans up most of
the code and "hopefully" makes it easier to follow and understand with less
work.