fs get updates #58
updated the fs command args to be more of what I tend to do. by default
it will attempt to retrieve the contents of a directory if the item
specified is a container of some kind. removed the roots command as if
the id is not present then it will just do that instead of an id lookup.
small change to sorting.