updated sec roles #58
updated the sec roles get commands. also updated the api structs by
adding some ordering for the permission, scope, and ability since they
are simple in what they represent. also added an additional function to
TextRow for using the ordering traits if the object in question
implements it as well.

fixed a bug in one of the requests that was not returning the pagination
data with the response.