changed top level arguments #58
removed the hash command as it is out of place with the rest of the cli
and all it does is the blake3 hash which is available as a executable on
their repo. also removed the interactive option as it is not necessary
and is not being used.
updated the host argument to be required and will do some parsing on the
given value to see if it contains the port number as well.
changed the secure argument to insecure and the client will default to
https vs http.