upload with hash #59
added in the ability to upload a hash of the file to the server to
validate against when creating new files. currently only supports blake3
hashes but I might add more later. the cli also handles sending hashes
to the server when requested.

some of the fs arguments were also updated to remove the need for some
of the options since they had to be specified reguardless so now they
are apart of command itself and not just options.

the config was updated to require a tmp directory for the upload
process. since the hash validation requires the file to be uploaded we
cannot write to the previous version if validation failed or something
else goes wrong. now files will be written to the tmp directory with the
fs id and then moved to the appropriate location. there can be a problem
with the file existing in tmp from a previous upload if something goes
wrong during the upload process. how to solve that I am not sure so for
now it will be an issue.