net module and serve_file
went through the net module and moved/removed code. net::error is no
longer in use. net::fs had the parts in use moved to serve_file since
that was the only thing using the code. net::mime code was moved to path
and html was just removed since there is no html to serve that is
generated by the server.