tmp file changes #60
have a new process for uploading files to the server and cleaning up tmp
files during the process. new file uploads are the simplest to handle
since only 1 file is created and handled. uploading an existing file
involves 2 files with a way to recover the previous file if something
goes wrong. moved the upload logic to its own file and separated out
more of the logic into separate functions just to try and clean things
up a bit. left some comments for the existing upload process to indicate
why the server is doing what its doing.