rbac cache #61
after some quick testing the time we save when caching permissions for a
user was noticable for response times. saving ~2ms on the server I am
using. updated a few spots with the new call and also updated the
necessary positions for clearing the cache when needed. having looked at
some of the code for users, groups, and roles there are parts that could
be cleaned up and moved into a single api path, mainly roles and groups.
also trying out using extractors more to get common things and not
having to directly access the state for every request when all we need
is the database and rbac cache.